Rename package group to `com.hhandoko`
