+++ examples/play25-example/app/com/hhandoko/ApplicationModule.scala
@@ -1,5 +1,5 @@
 /**
- * File     : HomeController.scala
+ * File     : ApplicationModule.scala
  * License  :
  *   The MIT License (MIT)
  *
@@ -24,46 +24,31 @@
  *   OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  *   SOFTWARE.
  */
-package com.builtamont.controllers
+package com.hhandoko
 
-import javax.inject._
-import scala.concurrent.ExecutionContext
+import com.google.inject.{AbstractModule, Provides}
+import net.codingwell.scalaguice.ScalaModule
 
-import play.api.Configuration
-import play.api.mvc._
+import _root_.play.api.Environment
 
-import com.builtamont.play.pdf.PdfGenerator
+import com.hhandoko.play.pdf.PdfGenerator
 
-/**
- * Home controller.
- *
- * @param config the application configuration
- * @param pdfGen the PDF generator implementation.
- */
-@Singleton
-class HomeController @Inject() (config: Configuration, pdfGen: PdfGenerator)(implicit exec: ExecutionContext) extends Controller {
+class ApplicationModule extends AbstractModule with ScalaModule {
 
-  val BASE_URL = config.getString("application.base_url").getOrElse("http://localhost:9000")
+  /** Module configuration + binding */
+  def configure(): Unit = {}
 
   /**
-   * Returns the homepage ('/').
+   * Provides PDF generator implementation.
    *
-   * @return Homepage.
+   * @param env The current Play app Environment context.
+   * @return PDF generator implementation.
    */
-  def index = Action { Ok(views.html.index()) }
-
-  /**
-   * Returns the example page as HTML ('/example').
-   *
-   * @return Example page.
-   */
-  def exampleHtml = Action { Ok(views.html.example() )}
-
-  /**
-   * Returns the example page as PDF document ('/example.pdf').
-   *
-   * @return Example page as PDF.
-   */
-  def examplePdf = Action { pdfGen.ok(views.html.example(), BASE_URL) }
+  @Provides
+  def providePdfGenerator(env: Environment): PdfGenerator = {
+    val pdfGen = new PdfGenerator(env)
+    pdfGen.loadLocalFonts(Seq("fonts/opensans-regular.ttf"))
+    pdfGen
+  }
 
 }
