+++ examples/play25-example/app/views/main.scala.html
@@ -62,13 +62,13 @@
                     <span class="icon-bar"></span>
                     <span class="icon-bar"></span>
                 </button>
-                <a class="navbar-brand" href="@com.builtamont.controllers.routes.HomeController.index().url">`play2-scala-pdf` Example Application</a>
+                <a class="navbar-brand" href="@com.hhandoko.controllers.routes.HomeController.index().url">`play2-scala-pdf` Example Application</a>
             </div>
             <div id="navbar" class="navbar-collapse collapse">
                 <ul class="nav navbar-nav">
-                    <li><a href="@com.builtamont.controllers.routes.HomeController.index().url">Home</a></li>
-                    <li><a href="@com.builtamont.controllers.routes.HomeController.exampleHtml().url">HTML Rendering</a></li>
-                    <li><a href="@com.builtamont.controllers.routes.HomeController.examplePdf().url">PDF Rendering</a></li>
+                    <li><a href="@com.hhandoko.controllers.routes.HomeController.index().url">Home</a></li>
+                    <li><a href="@com.hhandoko.controllers.routes.HomeController.exampleHtml().url">HTML Rendering</a></li>
+                    <li><a href="@com.hhandoko.controllers.routes.HomeController.examplePdf().url">PDF Rendering</a></li>
                 </ul>
             </div>
         </div>
