+++ examples/play25-example/conf/application.conf
@@ -39,7 +39,7 @@ play {
 
   # Register the user modules
   modules {
-    enabled += "com.builtamont.ApplicationModule"
+    enabled += "com.hhandoko.ApplicationModule"
   }
 
 }
