+++ examples/play26-example/app/views/index.scala.html
@@ -33,7 +33,7 @@
         `play2-scala-pdf` is a Play! module to help generate PDF documents dynamically from Play! web application.
         It simply renders Play! HTML and CSS-based view templates to PDF via <a href="https://github.com/flyingsaucerproject/flyingsaucer">Flying Saucer library</a>, which uses older, open-source version of iText for PDF generation.
     </p>
-    <p><a class="btn btn-primary btn-lg" href="https://github.com/builtamont/play2-scala-pdf" role="button">Github</a></p>
+    <p><a class="btn btn-primary btn-lg" href="https://github.com/hhandoko/play2-scala-pdf" role="button">Github</a></p>
 
 }
