+++ examples/play26-example/conf/application.conf
@@ -42,7 +42,7 @@ play {
 
   # Register the user modules
   modules {
-    enabled += "com.builtamont.ApplicationModule"
+    enabled += "com.hhandoko.ApplicationModule"
   }
 
   # Disable all filters for this example
