+++ examples/play26-example/conf/routes
@@ -35,7 +35,7 @@
 GET     /assets/*file               controllers.Assets.versioned(path="/public", file: Asset)
 
 # Pages
-GET     /example.pdf                com.builtamont.controllers.HomeController.examplePdf
-GET     /example                    com.builtamont.controllers.HomeController.exampleHtml
-GET     /                           com.builtamont.controllers.HomeController.index
+GET     /example.pdf                com.hhandoko.controllers.HomeController.examplePdf
+GET     /example                    com.hhandoko.controllers.HomeController.exampleHtml
+GET     /                           com.hhandoko.controllers.HomeController.index
