+++ modules/play24/app/com/hhandoko/play/pdf/PdfGenerator.scala
@@ -27,7 +27,7 @@
  * Notes:
  *   Also based on https://github.com/sihil/sihilsia/blob/master/app/controllers/pdf.scala
  */
-package com.builtamont.play.pdf
+package com.hhandoko.play.pdf
 
 import java.io._
 import scala.collection.mutable.ArrayBuffer
@@ -39,18 +39,17 @@ import org.apache.commons.io.{FilenameUtils, IOUtils}
 import org.w3c.tidy.Tidy
 import org.xhtmlrenderer.pdf.ITextRenderer
 
-import play.api.Environment
+import play.api.Play
 import play.api.mvc.{Result, Results}
 import play.twirl.api.Html
 
 /**
  * PDF generator service.
  *
- * @param env The current Play app Environment context.
  * @param xhtml true to set XHTML strict parsing (i.e. leave disabled for HTML5 templates).
  */
 @Singleton
-class PdfGenerator(env: Environment, val xhtml: Boolean = false) {
+class PdfGenerator(val xhtml: Boolean = false) {
 
   /** HTML tidy checker / prettyprinter instance for XHTML strict parsing */
   private lazy val tidyParser = {
@@ -80,7 +79,7 @@ class PdfGenerator(env: Environment, val xhtml: Boolean = false) {
    */
   def addTemporaryFonts(fonts: Seq[String]): Unit = {
     fonts.foreach { font =>
-      val stream = env.resourceAsStream(font)
+      val stream = Play.current.resourceAsStream(font)
 
       stream.map { s =>
         val file = File.createTempFile("tmp_" + FilenameUtils.getBaseName(font), "." + FilenameUtils.getExtension(font))
@@ -166,7 +165,7 @@ class PdfGenerator(env: Environment, val xhtml: Boolean = false) {
     val input = new ByteArrayInputStream(string.getBytes("UTF-8"))
     val renderer = new ITextRenderer()
     fonts.foreach { font => renderer.getFontResolver.addFont(font, BaseFont.IDENTITY_H, BaseFont.EMBEDDED) }
-    val userAgent = new PdfUserAgent(env, renderer.getOutputDevice)
+    val userAgent = new PdfUserAgent(renderer.getOutputDevice)
     userAgent.setSharedContext(renderer.getSharedContext)
     renderer.getSharedContext.setUserAgentCallback(userAgent)
     val document = new HtmlDocumentBuilder().parse(input)
