+++ modules/play25/publish.sbt
@@ -28,7 +28,7 @@ publishMavenStyle := true
 
 pomIncludeRepository := { _ => false }
 
-sonatypeProfileName := "com.builtamont"
+sonatypeProfileName := "com.hhandoko"
 
 publishTo := {
   val nexus = "https://oss.sonatype.org/"
@@ -39,17 +39,17 @@ publishTo := {
 }
 
 pomExtra in Global := (
-  <url>https://github.com/builtamont-oss/play2-scala-pdf</url>
+  <url>https://github.com/hhandoko/play2-scala-pdf</url>
   <licenses>
     <license>
       <name>MIT</name>
-      <url>https://raw.githubusercontent.com/builtamont-oss/play2-scala-pdf/master/LICENSE</url>
+      <url>https://raw.githubusercontent.com/hhandoko/play2-scala-pdf/master/LICENSE</url>
       <distribution>repo</distribution>
     </license>
   </licenses>
   <scm>
-    <url>git@github.com:builtamont-oss/play2-scala-pdf.git</url>
-    <connection>scm:git:git@github.com:builtamont-oss/play2-scala-pdf.git</connection>
+    <url>git@github.com:hhandoko/play2-scala-pdf.git</url>
+    <connection>scm:git:git@github.com:hhandoko/play2-scala-pdf.git</connection>
   </scm>
   <developers>
     <developer>
